PycharmProjects/SupremeCourtProject/requirements.txt
openai
tiktoken
pandas
transformers
plotly
matplotlib
sklearn
torch
torchvision
scipy
plotly
scipy
python-dotenv
flask_executor
slack_sdk
openai
numpy
pandas